\name{CPI975}
\alias{CPI975}
\title{Plasma levels of CPI975 in rats following a single oral dose}
\description{
  The \code{CPI975} data frame has 60 rows and 4 columns. 
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{conc}{
      a numeric vector giving the measured plasma level (microgram/mL).
    }
    \item{time}{
      a numeric vector giving the time since administration (hours).
    }
    \item{sex}{
      a factor indicating the gender of the rat.
    }
    \item{dose}{
      a numeric vector indicating the dose administred (mg/kg).
    }
  }
}
\details{
  Nedelman \emph{et al.} (1995), Table 2 describes serial sampling design were the plasma levels are measured at 5 time points over 24 hours for 10 rats. The same experiment was repeated with different oral dosages and gender.\cr
}
\source{
 Nedelman J. R., Gibiansky E. and Lau D. T. W. (1995). Applying Bailer's method for AUC confidence intervals to sparse sampling. \emph{Pharmaceutical Research}, 12(1):124-128. \cr\cr
}
%\examples{}
\keyword{datasets}
