\name{PKNews}
\alias{PKNews}
\title{Shows changes and news}
\description{Functions showing changes since previous versions.}
\usage{
PKNews()
}

\details{
Displays the changes and news given in the NEWS file of the package.
}

\value{Screen output.}

\author{Thomas Jaki}

\examples{
PKNews()
}

\keyword{misc}
