\name{plot.halflife}
\alias{plot.halflife}
\title{Plot Regression Lines Used for Half-life Estimation}

\description{
This method plots objects of S3 class \code{halflife}.
}

\usage{\method{plot}{halflife}(x, xlab='Time', ylab='Concentration', main='Half-life Estimation', xlim=NULL, ylim=NULL, ...)}

\arguments{
  \item{x}{ an object of S4 class link{\code{halflife}}}
  \item{xlab}{ a label for the x axis.}
  \item{ylab}{ a label for the y axis.}
  \item{main}{ a main title for the plot.}
  \item{xlim}{ the x limits (min, max) of the plot. }
  \item{ylim}{ the y limits (min, max) of the plot. }
  \item{\dots}{ other parameters to be passed through to plotting functions.}
}

\value{none}
\keyword{hplot}

