\name{Rats}
\alias{Rats}
\title{Plasma levels in female rats following a single oral dose}
\description{
  The \code{Rats} data frame has 126 rows and 4 columns. 
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{id}{
      a numeric vector identifying the animal whose plasma level is measured.
    }
    \item{conc}{
      a numeric vector giving the measured plasma level (microgram/mL).
    }
    \item{time}{
      a numeric vector giving the time since administration (hours).
    }
    \item{dose}{
      a numeric vector indicating the dose administred (mg/kg).
    }
  }
}
\details{
  Holder \emph{et al.} (1999), Table 4 describes data on the plasma levels measured at 6 time points over 24 hours for 9
  female rats. The same experiment was repeated with different oral dosages.\cr
}
\source{
  Holder D. J., Hsuan F., Dixit R. and Soper K. (1999). A method for estimating and testing area under the curve in serial sacrifice, batch, and complete data designs. \emph{Journal of Biopharmaceutical Statistics}, 9(3):451-464.
}
%\examples{}
\keyword{datasets}
